requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.3
coverage==5.0.4
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
pyyaml
six==1.14.0
sphinx==2.4.4
wcwidth==0.1.9
